requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.3
pymongo==4.6.0
redis==5.0.1
celery==5.3.4
polars==0.20.2
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
aiohttp==3.9.1
motor==3.3.2
dnspython==2.4.2
